rapi how sama do

+++ src/repositories/DoRepository.ts
@@ -1,47 +1,56 @@
 import { DoModel } from '../models/DoModel';
-import {HowModel} from "../models/HowModel";
 
 export class DoRepository {
-    async findAll() {
-        return DoModel.find()
+    async findAllByUser(userId: string) {
+        return DoModel.find({ createdBy: userId }).sort({ createdAt: -1 });
+    }
+
+    async findAllWithPopulateByUser(userId: string) {
+        return DoModel.find({ createdBy: userId })
+            .populate('nama_program')
             .sort({ createdAt: -1 });
     }
 
-    async findAllWithPopulate() {
-        return DoModel.find()
-            .sort({ createdAt: -1 });
+    async findById(id: string, userId: string) {
+        return DoModel.findOne({ _id: id, createdBy: userId }).populate('nama_program');
     }
 
-    async findById(id: string) {
-        return DoModel.findById(id);
+    async create(data: any, userId: string) {
+        return DoModel.create({ ...data, createdBy: userId });
     }
 
-    async create(data: any) {
-        // const program = new DoModel(data);
-        // return program.save();
-        return DoModel.create(data)
+    async update(id: string, data: any, userId: string) {
+        return DoModel.findOneAndUpdate(
+            { _id: id, createdBy: userId },
+            data,
+            { new: true }
+        );
     }
 
-    async update(id: string, data: any) {
-        return DoModel.findByIdAndUpdate(id, data, { new: true });
+    async delete(id: string, userId: string) {
+        return DoModel.findOneAndDelete({ _id: id, createdBy: userId });
     }
 
-    async delete(id: string) {
-        return DoModel.findOneAndDelete({ id });
-    }
-
-    async search(query: string) {
+    async search(query: string, userId: string) {
         return DoModel.find({
+            createdBy: userId,
             $or: [
-                { nama_program: { $regex: query, $options: 'i' } },
                 { rincian_kegiatan: { $regex: query, $options: 'i' } },
                 { capaian_output: { $regex: query, $options: 'i' } },
                 { dokumentasi_kegiatan: { $regex: query, $options: 'i' } },
                 { kendala: { $regex: query, $options: 'i' } },
                 { rekomendasi: { $regex: query, $options: 'i' } },
-                { kolaborator: { $elemMatch: { nama: { $regex: query, $options: 'i' } } } },
-                { kolaborator: { $elemMatch: { peran: { $regex: query, $options: 'i' } } } },
+                {
+                    kolaborator: {
+                        $elemMatch: {
+                            $or: [
+                                { nama: { $regex: query, $options: 'i' } },
+                                { peran: { $regex: query, $options: 'i' } }
+                            ]
+                        }
+                    }
+                }
             ]
-        }).sort({ createdAt: -1 });
+        }).populate('nama_program').sort({ createdAt: -1 });
     }
 }
